+++ ethereum/spec/src/test/java/tech/pegasys/teku/spec/SpecFactoryTest.java
@@ -40,9 +40,9 @@ public class SpecFactoryTest {
   private static final Set<String> NON_BELLATRIX_NETWORKS = Set.of("swift", "less-swift");
 
   private static final Set<String> CAPELLA_NETWORKS =
-      Set.of("sepolia", "mainnet", "gnosis", "lukso", "holesky");
+      Set.of("mainnet", "gnosis", "lukso");
 
-  private static final Set<String> DENEB_NETWORKS = Set.of("prater", "chiado");
+  private static final Set<String> DENEB_NETWORKS = Set.of("prater", "chiado", "sepolia", "holesky");
 
   @Test
   public void defaultFactoryShouldScheduleBellatrixAndCapellaForMainNet() {